Move views to v2 subfolder
