+++ client/src/views/v2/CameraSettingsView.js
