+++ client/src/views/v2/PowerMapView.js
