+++ client/src/views/v2/SensorStatusView.js
